requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
SQLAlchemy==2.0.28
Werkzeug==3.0.1
python-dotenv==1.0.1
requests-oauthlib==1.3.1
Flask-Migrate==4.0.7